Movendo estado da aba selecionada para o contexto do sistema
