+++ src/context/System/systemContext.tsx
@@ -1,10 +1,14 @@
-import React, { useCallback, useContext, useState } from "react";
+import React, { useCallback, useContext, useMemo, useState } from "react";
 
 interface SystemContextData {
   isOpenRightSideMenu: boolean;
   changeRightSideMenuVisibility: (visibility?: boolean) => void;
   activeAppTab: ActiveAppTab;
   setActiveAppTab: React.Dispatch<React.SetStateAction<ActiveAppTab>>;
+  handleTabChange: (key: string | null) => void;
+  selectedTab: string;
+  selectedTabIndex: number;
+  setSelectedTab: React.Dispatch<React.SetStateAction<string>>;
 }
 
 const SystemContext = React.createContext<SystemContextData>(
@@ -17,6 +21,23 @@ const SystemProvider: React.FC = ({ children }) => {
   const [isOpenRightSideMenu, setIsOpenRightSideMenu] = useState(false);
   const [activeAppTab, setActiveAppTab] = useState<ActiveAppTab>("plans");
 
+  const [selectedTab, setSelectedTab] = useState<string>(() => {
+    const tabKey = localStorage.getItem("@plan:selectedTab");
+
+    if (tabKey) {
+      return tabKey;
+    }
+
+    return "tab1";
+  });
+
+  const handleTabChange = useCallback((key: string | null) => {
+    if (key) {
+      localStorage.setItem("@plan:selectedTab", key);
+      setSelectedTab(key);
+    }
+  }, []);
+
   const changeRightSideMenuVisibility = useCallback((visibility?: boolean) => {
     if (visibility === true || visibility === false) {
       setIsOpenRightSideMenu(visibility);
@@ -26,6 +47,10 @@ const SystemProvider: React.FC = ({ children }) => {
     }
   }, []);
 
+  const selectedTabIndex = useMemo(() => {
+    return Number(selectedTab.replace("tab", ""));
+  }, [selectedTab]);
+
   return (
     <SystemContext.Provider
       value={{
@@ -33,6 +58,10 @@ const SystemProvider: React.FC = ({ children }) => {
         changeRightSideMenuVisibility,
         activeAppTab,
         setActiveAppTab,
+        handleTabChange,
+        selectedTab,
+        selectedTabIndex,
+        setSelectedTab,
       }}
     >
       {children}
